% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Control_probe_chain_file}
\alias{Control_probe_chain_file}
\title{The list of control probes in Illumina methylation BeadChip}
\format{
A data frame with 985 rows and 8 columns
}
\usage{
data(Control_probe_chain_file)
}
\description{
This file contains control probes in EPICv2, EPICv1, and 450K.
This file generated based on the Illumina manifests from 3 types of platform.

This file is used for calculating detection P-value and normalization step.
}
\keyword{datasets}
