% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Duplicated.Probes.preset}
\alias{Duplicated.Probes.preset}
\title{The preset data of duplicated probe to remove in EPICv2}
\format{
A data frame with 6963 rows and 28 columns
}
\usage{
data(Duplicated.Probes.preset)
}
\description{
This file contains probe IDs of duplicated probe in EPICv2.
Using this file, the address chain file can be generated, which is a
reference file to analysis EPICv2 data or convert the EPICv2 data
into previous versions.

To customize this file, we recommend to use Illumina manifest file.
(custom <- IlluminaHumanMethylationEPIC.V2a1.hg38\link{removelist, })

Please refer to this file to create customized data for the \code{MeCall.SetChainFile} function.

This file is used to address duplicated probe and generate \code{address_chain_file} for preprocessing of EPICv2.
}
\keyword{datasets}
