% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EPIC.MultiHit.Nordlund_GenomeBiology_2013}
\alias{EPIC.MultiHit.Nordlund_GenomeBiology_2013}
\title{Multi-Hit probe list in EPICv1}
\format{
A data frame with 9341 rows and 37 columns
}
\usage{
data(EPIC.MultiHit.Nordlund_GenomeBiology_2013)
}
\description{
This file contains the probe list that have multiple aligned sequences.
This file can be found in supplementary materials of Nordlund's study
(doi : url{https://doi.org/10.1186/gb-2013-14-9-r105}).

This file is used for filtering probes.
}
\references{
Nordlund, J. et al. (2013). Genome-wide signatures of differential DNA methylation in pediatric acute lymphoblastic leukemia. Genome Biology. url{https://doi.org/10.1186/gb-2013-14-9-r105}
}
\keyword{datasets}
