% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FLAGprobes}
\alias{FLAGprobes}
\title{Flagged probe list}
\format{
A data frame with 50209 rows and 48 columns
}
\usage{
data(FLAGprobes)
}
\description{
This file is a subset of the Illumina EPICv2 manifest that
includes flagged probes.
This file is included in the Infinium MethylationEPIC v2.0 Product Files
from Illumina official website (\url{https://support.illumina.com/array/array_kits/infinium-methylationepic-beadchip-kit/downloads.html}).

This file is used for filtering probes.
}
\keyword{datasets}
