% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Illumina_manifest}
\alias{Illumina_manifest}
\alias{IlluminaHumanMethylationEPIC.V2a1.hg38}
\alias{IlluminaHumanMethylationEPIC.V2a1.hg38,}
\alias{IlluminaHumanMethylationEPIC.V1b5.hg38,}
\alias{IlluminaHumanMethylation450K}
\title{Illmunima manifest files}
\format{
A data frame with 937691 rows and 28 columns (EPICv2)
}
\usage{
data(IlluminaHumanMethylationEPIC.V2a1.hg38)
}
\description{
These files are the Illumina manifest file downloaded from
the Illumina official website (\url{https://support.illumina.com/?tab=microarrays}).
To reduce the file size, several columns have been removed.
The file header and information of control probes have also been deleted.
The information of control probes can be found in \code{?Control_probe_chain_file}.

These files are used for annotation of probes.
}
\keyword{datasets}
