% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapping_failed_probe}
\alias{mapping_failed_probe}
\alias{V2a1.hg38.mapping.fail}
\alias{V2a1.hg38.mapping.fail,}
\alias{V1b5.hg38.mapping.fail,}
\alias{hg38.mapping.fail.450K}
\alias{V1b5.hg38.mapping.fail}
\title{Mapping failed probe list}
\format{
A data frame with 190 rows and 48 columns (EPICv2)

A data frame with 38 rows and 51 columns (EPICv1)

A data frame with 32 rows and 27 columns (450K)
}
\usage{
data(V2a1.hg38.mapping.fail)

data(V1b5.hg38.mapping.fail)

data(hg38.mapping.fail.450K)
}
\description{
These files are a subset of the Illumina manifest or
SeSaMe manifest that include mapping failed probes.

\code{V2a1.hg38.mapping.fail} for EPICv2 (Illumina)
\code{V1b5.hg38.mapping.fail} for EPICv1 (Illumina)
\code{hg38.mapping.fail.450K} for 450K (SeSaMe)

Each row represents a probe.
Documentation for each column can be found in the Illumina product files (\url{https://support.illumina.com/array/array_kits/infinium-methylationepic-beadchip-kit/downloads.html})
or github pages of SeSaMe (\url{https://zwdzwd.github.io/InfiniumAnnotation})

These files are used for filtering probes.
}
\references{
For 450K -
Zhou, W. et al. (2017). Comprehensive characterization, annotation and innovative use of Infinium DNA methylation BeadChip probes. Nucleic Acids Research. \url{https://doi.org/10.1093/nar/gkw967}
}
\keyword{datasets}
