% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.CellComp.R
\name{MeCall.CellComp}
\alias{MeCall.CellComp}
\title{Estimate cell composition of various tissue type.}
\usage{
MeCall.CellComp(
  data = Filtered,
  meth = beta,
  processMethod = "preprocessNoob",
  CellType = "Blood",
  method = c("EpiSCORE", "deconvR", "manual"),
  Reference = NULL,
  adjust = TRUE,
  return.All = FALSE,
  arraytype = "EPICv1"
)
}
\arguments{
\item{data}{A return object from \code{MeCall.Filtering()} for blood tissue type.}

\item{meth}{A matrix of methylation level for solid tissue type or adjustment.}

\item{processMethod}{Option for blood tissue type. The normalization method to be applied along with the reference data ("preprocessNoob", "preprocessFunnorm", "preprocessQuantile", "preprocessSwan", "preprocessRaw").}

\item{CellType}{The tissue type of the sample.}

\item{method}{Option for solid tissue type. The reference matrix of the selected method will be applied.}

\item{Reference}{Data frame of reference to be used for cell type estimation. The columns are the sample names, and the rows are the CpG IDs. The \code{method} parameter must be set to \code{manual}.}

\item{adjust}{Logical. Methylation levels should be provided. Methylation levels are adjusted and returned based on the estimated cell composition.}

\item{return.All}{Logical. Both the adjusted methylation levels and a data frame of the estimated cell type composition are returned.}

\item{arraytype}{The version of the microarray.}
}
\value{
A data frame of estimated cell compostion or a matrix of adjusted methylation level. If \code{return.All} is set to TRUE, a list object containing two data frames will be returned.
}
\description{
This function involves estimating the proportions of cell types for each
sample based on a selected or provided reference matrix and correcting
biases using linear regression that result from cell type heterogeneity
between samples.

MethylCallR provides a function to estimate white blood cell composition
using the \code{FlowSorted.Blood.EPIC} R packages.

In case of solid tissue samples, estimation of the cell type proportions is
conducted using reference matrices and functions implemented in
the \code{EPISCORE} and \code{deconvR} R packages.

If a reference matrix is available, the user may want to set the \code{method}
parameter to \code{manual} and input the reference matrix into the \code{Reference}
parameter. Using the \code{deconvolute()} function implemented in the \code{deconvR} R
package, the cell type proportions will be calculated. Please ensure that
the order of probes in the reference matrix matches the order of probes
in the methylation level matrix.

If the \code{adjust} parameter is set to \code{TRUE}, \code{MethylCallR} will automatically
correct for estimated cell type heterogeneity using linear regression for
all inferred cell types. This function consecutively performs estimation
and correction, and if \code{return.All} is also set to \code{TRUE}, the estimated
cell type proportions and the corrected methylation levels will be
returned together as part of the result object.
}
\examples{
\dontrun{
# Run [MeCall.CellComp] with Blood tissue type.
data.CellComp <- MeCall.CellComp(data = data.filtered, meth = data.rmBatch$methylation.level, 
processMethod = "preprocessNoob", CellType = "Blood", adjust = TRUE, return.All = FALSE)

# Run [MeCall.CellComp] with Solid tissue type.
data.CellComp <- MeCall.CellComp(meth = data.rmBatch$methylation.level, CellType = "Lung", 
method = c("EpiSCORE"), adjust = TRUE, return.All = FALSE,  arraytype = "EPICv1")
}

}
\references{
Adjustment code comes from
Jones, M.J. et al. (2017). Adjusting for Cell Type Composition in DNA Methylation Data Using a
Regression-Based Approach. Methods in Molecular Biology (Clifton, N.J.).
/url{https://doi.org/10.1007/7651_2015_262}

For FlowSorted.Blood.EPIC
Salas, L.A. et al. (2018). An optimized library for reference-based deconvolution of
whole-blood biospecimens assayed using the Illumina HumanMethylationEPIC BeadArray.
Genome Biology. /url{https://doi.org/10.1186/s13059-018-1448-7}

For EPISCORE
Teschendorff, A.E. et al. (2020). EPISCORE: cell type deconvolution of bulk tissue DNA
methylomes from single-cell RNA-Seq data. Genome Biology.
/url{https://doi.org/10.1186/s13059-020-02126-9}

For deconvR
Moss, J. et al. (2018). Comprehensive human cell-type methylation atlas reveals origins of
circulating cell-free DNA in health and disease. Nature Communications.
/url{https://doi.org/10.1038/s41467-018-07466-6}
}
\seealso{
\code{\link[FlowSorted.Blood.EPIC:estimateCellCounts2]{FlowSorted.Blood.EPIC::estimateCellCounts2()}},  \code{\link[deconvR:deconvolute]{deconvR::deconvolute()}}
}
\author{
Hyun-Ho Yang
}
