% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.CNV.R
\name{MeCall.CNV}
\alias{MeCall.CNV}
\title{Perform the copy number variation (CNV) analysis}
\usage{
MeCall.CNV(
  data = data.filtered,
  interest = "Sample_Group",
  batch = c("Slide", "Array"),
  arraytype = "EPIC",
  outdir = NULL,
  intercept = TRUE
)
}
\arguments{
\item{data}{A return object from \code{MeCall.filtering()}.}

\item{interest}{The column name in the sample data file containing the phenotype of interest.}

\item{batch}{Before estimating CNVs, the given batch effect will be corrected using \code{ComBat} algoritm. The variable name must be included in the column names of the sample data file.}

\item{arraytype}{Version of the methylation microarray.}

\item{outdir}{Directory path to save the results in \code{GISTIC2.0} input format.}

\item{intercept}{logical. If set to \code{TRUE}, the intercept will also be considered. (for regression method)}
}
\value{
A list object containing the segment results per sample and the marker information.
}
\description{
Copy number variation between two groups is identified using intensity
values. \code{MethylCallR} provides an automated function that performs
data normalization, technical batch effect adjustment, and segmentation.
Finally, \code{MeCall.CNV()} returns a list object containing input files
for the \code{GISTIC2.0} algorithm, which is used to identify significantly
recurring focal alterations. By entering a directory in the
\code{outdir} parameter, the function assumes that the user agrees to
automatic file creation and saves the Segment and Marker files
in the specified directory :
\itemize{
\item Segment file : \verb{/outdir/MethylCallR_GISTIC_Segment_file.txt}
\item Marker file : \verb{/outdir/MethylCallR_GISTIC_Marker_file.txt}
}
}
\examples{
\dontrun{
# Run [MeCall.CNV] with default setting.
data.CNV <- MeCall.CNV(data = data.filtered, interest = "Sample_Group",
batch = c("Slide","Array"), arraytype=c("EPICv1"), outdir="./", intercept = TRUE)
}

}
\references{
Venkatraman, E.S. and Olshen, A.B. (2007). A faster circular binary segmentation algorithm
for the analysis of array CGH data. Bioinformatics.
/url{https://doi.org/10.1093/bioinformatics/btl646}

For fitting method
Tian, Y. et al. (2017). ChAMP: updated methylation analysis pipeline for Illumina BeadChips.
Bioinformatics. /url{https://doi.org/10.1093/bioinformatics/btx513}
}
\seealso{
\code{\link[ChAMP:champ.CNA]{ChAMP::champ.CNA()}}, \code{\link[DNAcopy:CNA]{DNAcopy::CNA()}}
}
\author{
Hyun-Ho Yang
}
