% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.DMP.R
\name{MeCall.DMP}
\alias{MeCall.DMP}
\title{Perform differential methylation analysis to identify differentially methylated probes (DMPs).}
\usage{
MeCall.DMP(
  meth = MeCall.CellComp$adjusted.meth,
  pd = MeCall.RemoveBatch$modified.pd,
  interest = "Sample_Group",
  design = data.design,
  cutoff.P = 1,
  multi.P = "BH",
  arraytype = "EPICv1"
)
}
\arguments{
\item{meth}{A matrix of methylation level.}

\item{pd}{A data frame of modified sample data file from \code{MeCall.RemoveBatch()}.}

\item{interest}{The column names in the sample data file containing the phenotype of interest.}

\item{design}{A design matrix from \code{MeCall.MakeModelMatrix()}}

\item{cutoff.P}{Probes with p-value above the specified threshold, as determined by the method specified in the \code{multi.P} parameter, are excluded from the results. If set to \code{1}, you can obtain results for all probes.}

\item{multi.P}{Set the method for multiple testing correction ("none","BH","BF").}

\item{arraytype}{Obtain annotation information for the given array type.}
}
\value{
"DMPresultlist" S4 class object consisted of :
\itemize{Main.info} {The data frame contains statistical values and major information for each probe. This information includes chromosome, position, functional region, CGI region, and gene name. For a detailed information of the statistical values, refer to the \code{\link[limma:toptable]{limma::topTable()}} function.}
\itemize{Other.info} {The data frame contains supplementary details from the Illumina manifest, excluding the major information.}
\itemize{accessory} {The vector includes the array type, the method used for multiple testing correction, the p-value threshold, and the names of the comparison groups.}
}
\description{
Identify differentially methylated probes between two groups using
preprocessed methylation level data.

Only probes with p-values adjusted for multiple testing using
the \code{multi.P} method and less than or equal to \code{cutoff.P} are
included in the result object.

By specifying the appropriate \code{arraytype} parameter (\code{450L}, \code{EPICv1},
\code{EPICv2}), accurate positional information and characteristics for
each probe can be provided. MethylCallR provides positional information
based on the GRCh38/hg38 human reference as a default, while
the GRCh37/hg19 human reference is included as additional information.
}
\examples{
\dontrun{
# Run [MeCall.DMP] with default setting.
data.DMPs <- MeCall.DMP(meth = MeCall.CellComp$adjusted.meth, 
pd = MeCall.RemoveBatch$modified.pd, interest = "Sample_Group", design = data.design, 
cutoff.P=1, multi.P = 'BH', arraytype="EPICv1")
}

}
\references{
Ritchie, M.E. et al. (2015). limma powers differential expression analyses for RNA-sequencing
and microarray studies. Nucleic Acids Research. /url{https://doi.org/10.1093/nar/gkv007}
}
\seealso{
\code{\link[limma:toptable]{limma::topTable()}}
}
\author{
Hyun-Ho Yang
}
