% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.DMR.R
\name{MeCall.DMR}
\alias{MeCall.DMR}
\title{Perform differential methylation analysis to identify differentially methylated regions (DMRs).}
\usage{
MeCall.DMR(
  meth = data.rmBatch$methylation.level,
  pd = data.rmBatch$modified.pd,
  interest = "Sample_Group",
  covariate = NULL,
  arraytype = c("EPICv2", "EPICv1", "450K"),
  method = c("bumphunter", "seqlm", "dmrff", "DMRcate", "combP"),
  adjPvalDmr = 1,
  nProbes = 2,
  probe_gap = NULL,
  cutoff = NULL,
  pickCutoff = TRUE,
  pickCutoffQ = 0.95,
  smooth = TRUE,
  smoothFunction = loessByCluster,
  useWeights = FALSE,
  B = 100,
  nullMethod = "bootstrap",
  max_block_length = 50,
  dmp.fdr = 0.05,
  C = 5
)
}
\arguments{
\item{meth}{A matrix of methylation level.}

\item{pd}{A data frame of modified sample data file from \code{MeCall.RemoveBatch()}.}

\item{interest}{The column name in the sample data file containing the phenotype of interest.}

\item{covariate}{Column names of covariates to be considered when identifying differentially methylated regions.}

\item{arraytype}{Version of the methylation microarray.}

\item{method}{The name of the algorithm to be used for identifying differentially methylated regions.}

\item{adjPvalDmr}{The p-value threshold for identifying significant differentially methylated regions.}

\item{nProbes}{The minimum number of probes included within a significant differentially methylated region.}

\item{probe_gap}{The minimum distance between probes to be considered as different regions. If set to NULL, the recommended or default values for each algorithm will be automatically applied. (bumphunter : maxGap; seqlm : max_dist; dmrff : maxgap; DMRcate : lambda)}

\item{cutoff}{(Bumphunter) Threshold values to identify candidate regions. Upper and lower threshold values can be provided individually.}

\item{pickCutoff}{(Bumphunter) logical. If set to \code{TRUE}, an appropriate cutoff is selected using the permutation distribution.}

\item{pickCutoffQ}{(Bumphunter) The quantile value for the cutoff to be selected from the permutation distribution.}

\item{smooth}{(Bumphunter) logical. If set to \code{TRUE}, Smoothing will be performed to the estimated profile using the smoothFunction.}

\item{smoothFunction}{(Bumphunter) The name of the function to be applied for smoothing.}

\item{useWeights}{(Bumphunter) logical. If set to \code{TRUE} and smoothing function is set to \code{loessByCluster}, the standard errors of the point-wise estimates of the profile function will be used as weights.}

\item{B}{(Bumphunter) The number of resamples for generating the null distribution.}

\item{nullMethod}{(Bumphunter) The method for generating the null distribution.}

\item{max_block_length}{(seqlm) The maximum number of CpGs that can be included in one region.}

\item{dmp.fdr}{(DMRcate) The FDR threshold (Benjamini-Hochberg method) for identifying significant differential methylation probes.}

\item{C}{(DMRcate) Scaling factor for bandwidth. Lambda/C is defined as sigma, and a Gaussian kernel is computed accordingly.}
}
\value{
A data frame containing the genomic locations, associated genes,
and major statistical information of the identified DMR.
}
\description{
This function provides multiple useful algorithms for identifying
differentially methylated regions.

\code{MeCall.DMR()} function supports 5 algorithms :
\itemize{
\item Bumphunter (\code{method = "bumphunter"})
\item seqlm (\code{method = "seqlm"})
\item dmrff (\code{method = "dmrff"})
\item DMRcate (\code{method = "DMRcate"})
\item combined-pvalues (\code{method = "combP"})
}

Unlike other algorithms, \code{combined-pvalues} is a python-based algorithm.
Instead of returning DMR results, \code{MeCall.DMR()} will return the basic
file format (BED format) and a guide message required for using the
algorithm. For detailed instructions, please refer to GitHub :
https://github.com/brentp/combined-pvalues
}
\examples{
\dontrun{
# Run [MeCall.DMR] - Perform Bumphunter with default setting.
data.DMRs.Bump <- MeCall.DMR(meth = data.rmBatch$methylation.level, pd = data.rmBatch$modified.pd
, interest="Sample_Group", covariate = NULL, arraytype = c("EPICv2"), method = c("bumphunter"), 
adjPvalDmr=1, nProbes=2, probe_gap = 250, cutoff = NULL, pickCutoff = TRUE, pickCutoffQ = 0.95, 
smooth=TRUE, smoothFunction=loessByCluster, useWeights=FALSE, B=100, nullMethod="bootstrap")

# Run [MeCall.DMR] - Perform seqlm with default setting.
data.DMRs.seq <- MeCall.DMR(meth = data.rmBatch$methylation.level, pd = data.rmBatch$modified.pd,
interest="Sample_Group", covariate = NULL, arraytype = c("EPICv2"), method = c("seqlm"), 
adjPvalDmr=1, nProbes=2, probe_gap = 1000, max_block_length=50)

# Run [MeCall.DMR] - Perform dmrff with default setting.
data.DMRs.dmrff <- MeCall.DMR(meth = data.rmBatch$methylation.level, 
pd = data.rmBatch$modified.pd, interest="Sample_Group", covariate = NULL, 
arraytype = c("EPICv2"), method = c("dmrff"), adjPvalDmr=1, nProbes=2, probe_gap = 500)

# Run [MeCall.DMR] - Perform DMRcate with default setting.
data.DMRs.DMRcate <- MeCall.DMR(meth = data.rmBatch$methylation.level, 
pd = data.rmBatch$modified.pd, interest="Sample_Group", covariate = NULL, 
arraytype = c("EPICv2"), method = c("DMRcate"), adjPvalDmr=1, nProbes=2, probe_gap = 500, 
dmp.fdr=0.05, C=5)

# Run [MeCall.DMR] - Perform combP with default setting.
data.DMRs.combP.BEDfile <- MeCall.DMR(meth = data.rmBatch$methylation.level, 
pd = data.rmBatch$modified.pd, interest="Sample_Group", covariate = NULL, 
arraytype = c("EPICv2"), method = c("combP"))
}

}
\references{
Default setting of algorithms based on
Mallik, S. et al. (2019). An evaluation of supervised methods for identifying differentially
methylated regions in Illumina methylation arrays. Briefings in Bioinformatics.
/url{https://doi.org/10.1093/bib/bby085}

For bumphunter
Jaffe, A.E. et al. (2012). Bump hunting to identify differentially methylated regions in
epigenetic epidemiology studies. International Journal of Epidemiology.
/url{https://doi.org/10.1093/ije/dyr238}

For dmrff (BioRxiv)
Suderman, M. et al. (2018). dmrff: identifying differentially methylated regions efficiently
with power and control. /url{https://doi.org/10.1101/508556}

For DMRcate
Peters, T.J. et al. (2015). De novo identification of differentially methylated regions in the
human genome. Epigenetics & Chromatin. /url{https://doi.org/10.1186/1756-8935-8-6}

For seqlm
Kolde, R. et al. (2016). seqlm: an MDL based method for identifying differentially methylated
regions in high density methylation array data. Bioinformatics.
/url{https://doi.org/10.1093/bioinformatics/btw304}

For comb-P
Pedersen, B.S. et al. (2012). Comb-p: software for combining, analyzing, grouping and correcting
spatially correlated P-values. Bioinformatics. https://doi.org/10.1093/bioinformatics/bts545
}
\seealso{
\code{\link[bumphunter:bumphunter]{bumphunter::bumphunter()}}, \code{\link[DMRcate:dmrcate]{DMRcate::dmrcate()}}
}
\author{
Hyun-Ho Yang
}
