% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.Enrichment.R
\name{MeCall.Enrichment}
\alias{MeCall.Enrichment}
\title{Perfome functional enrichment analysis.}
\usage{
MeCall.Enrichment(
  DMP.object = data.DMPs,
  DMP.cutoff = 0.05,
  DMP.multi = "BH",
  Delta.Beta = c(-0.02, 0.02),
  genomic.features = c("ALL"),
  Enrich.cutoff = 0.05,
  Enrich.multi = c("none", "BF", "BH"),
  min.count = 2,
  minfi.object.EPICv2 = NULL
)
}
\arguments{
\item{DMP.object}{A returned object from \code{MeCall.DMP()}. Please note that the results should include information for all probes (set to \code{cutoff.P = 1}).}

\item{DMP.cutoff}{The p-value cutoff for DMPs to be included in functional enrichment analysis.}

\item{DMP.multi}{Filter DMPs using the p-values adjusted with the selected multiple testing method ("none","BF", "BH").}

\item{Delta.Beta}{Filter DMPs using the specified delta-beta cutoff.}

\item{genomic.features}{Functional enrichment analysis is performed using only CpG sites located within the restricted functional region.}

\item{Enrich.cutoff}{The p-value cutoff for filtering significant descriptions.}

\item{Enrich.multi}{Filter significant descriptions using the p-values adjusted with the selected multiple testing method ("none","BF", "BH").}

\item{min.count}{Consider results significant only if the description includes the selected number or more genes.}

\item{minfi.object.EPICv2}{To perform functional enrichment analysis using results from EPICv2, input the data object from \code{MeCall.ReadIdat()} function to obtain the list of genes associated with CpGs.}
}
\value{
A dataframe consisting of significantly enriched GO terms and KEGG pathways.
}
\description{
This function selects significant differentially methylated probes
and tests which gene sets from gene ontology and KEGG database are
significantly enriched based on these probes. Furthermore, the results
will be restructured to allow users to view all outcomes at once.

The \code{gometh()} function implemented in the \code{missMethyl} R package
performs functional enrichment analysis using Wallenius' noncentral
hypergeometric distribution to minimize biases resulting from
multiple relationships between CpG sites and genes.

When conducting functional enrichment analysis using EPICv2,
a matched manifest object is required due to its unique probe ID.
Since \code{MethylCallR} generates an \code{RGChannelSet} when loading IDAT files,
the annotation object required by \code{gometh()} could be easily created.
The \code{MeCall.enrichment()} function will automatically perform this
process by passing the return object of the \code{MeCall.enrichment()} function
to the \code{minfi.object.EPICv2} parameter.
}
\examples{
\dontrun{
# Run functional enrichment analysis with default setting.
data.Enrichment <- MeCall.Enrichment(DMP.object = data.DMPs, DMP.cutoff = 0.05, DMP.multi = "BH",
Delta.Beta = c(-0.02,0.02), genomic.features = c("ALL"), Enrich.cutoff = 0.05, 
Enrich.multi = c("none"),min.count = 2 ,minfi.object.EPICv2 = NULL)

# Run functional enrichment analysis using EPICv2 data.
data.Enrichment.EPICv2 <- MeCall.Enrichment(DMP.object = data.DMPs, DMP.cutoff = 0.05, 
DMP.multi = "BH", Delta.Beta = c(-0.02,0.02), genomic.features = c("ALL"), Enrich.cutoff = 0.05,
Enrich.multi = c("none"),min.count = 2 ,minfi.object.EPICv2 = data.Import)
}

}
\references{
Maksimovic, J. et al. (2021). Gene set enrichment analysis for genome-wide DNA methylation data.
Genome Biology. url{https://doi.org/10.1186/s13059-021-02388-x}
}
\seealso{
\code{\link[missMethyl:gometh]{missMethyl::gometh()}}
}
\author{
Hyun-Ho Yang
}
