% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.Filtering.R
\name{MeCall.Filtering}
\alias{MeCall.Filtering}
\title{Filtering samples and probes under various conditions for quality control}
\usage{
MeCall.Filtering(
  data = Idat.list,
  RemoveDup = TRUE,
  badSample.minfi = TRUE,
  badSampleCutoff.minfi = 10.5,
  detectionP = TRUE,
  detP.probe.cutoff = 0.01,
  detP.sample.cutoff = 0.1,
  detP.failed.ratio = 0,
  beadcount = TRUE,
  beadcutoff = 0.05,
  NoCG = TRUE,
  SNPmask = TRUE,
  population = "General",
  Multihit = TRUE,
  Multihit.method = "WGBS",
  Sexchr = "XY",
  FLAG = TRUE,
  hg38.mapfail = TRUE
)
}
\arguments{
\item{data}{A return object from \code{MeCall.ReadIdat()}}

\item{RemoveDup}{Option for EPICv2 data. When set to \code{MethylCallR}, code removes duplicated probes and generate \code{Address_chain_file} based on the data implemented in MethylCallR. If user provides customized data, code will remove duplicated probes based on the user data. If it set to \code{none}, code will pass this step.}

\item{badSample.minfi}{Logical. The sample filtering method implemented in the minfi R package is applied. The quality of each sample will be assessed based on the median of log-scaled intensity values.}

\item{badSampleCutoff.minfi}{Samples with values below the threshold are removed.}

\item{detectionP}{Logical. Filtering probes based on detection P-value matrix.}

\item{detP.probe.cutoff}{Probes with detection P-values above the threshold are removed.}

\item{detP.sample.cutoff}{Samples with a proportion of failed probes in detection P-value criteria above the threshold are removed.}

\item{detP.failed.ratio}{Detection P-value failed probes with a proportion above the threshold among the entire sample are removed. For example, if detP.failed.ratio = 0, any probe that fails the detection P-value criterion in at least one sample is removed from downstream analysis.}

\item{beadcount}{Logical. Filtering probes based on beadcount matrix.}

\item{beadcutoff}{Probes with a bead count of less than 3 are removed if the probe has a proportion above the threshold in the entire sample.}

\item{NoCG}{Logical. Probes that are not CG are removed.}

\item{SNPmask}{Logical. Probes located near SNPs of the selected population are removed}

\item{population}{SNPmask filtering is based on the selected population. ('General', 'AFR', 'EAS', 'EUR', 'SAS', 'AMR', 'GWD', 'YRI', 'TSI', 'IBS', 'CHS', 'PUR', 'JPT', 'GIH', 'CHASK', 'STU', 'ITU', 'LWK', 'KHV', 'FIN', 'ESN', 'CEU', 'PJL', 'ACASK', 'CLM', 'CDX', 'GBR', 'BEASK', 'PEL', 'MSL', 'MXL', 'ASW').}

\item{Multihit}{Logical. Probes with multiple aligned sequences will be removed.}

\item{Multihit.method}{Setting the criteria for removing cross-hybridizing probes. Use "WGBS" to remove probes that show stronger affinity for non-target sites identified by comparing to whole-genome bisulfite sequencing. Use "BLAT" to remove all probes with in silico cross-hybridization events predicted at least once by BLAT (default = "WGBS"). See (TJ Peters et al., 2024) for detail.}

\item{Sexchr}{Character. Probes existing on the selected sex chromosome are removed.}

\item{FLAG}{Logical. Filtering out probes with a low replication rate that does not meet the manufacturer's quality control criteria.}

\item{hg38.mapfail}{Logical. Filtering out probes with an ambiguous position in the GRCh38/hg38 reference genome.}
}
\value{
A list object containing basic informations with filtered CpG sites
to process methylation microarray. Return object of \code{MeCall.Filtering()} has
the similar structure as the \code{MeCall.ReadIdat()} function.
All filtered-out sample and probe names will be added to
\code{filtered.Sample.list} and \code{filtered.CpG.list}, respectively, and returned
together in the return object.
}
\description{
Removes samples and CpG sites with poor conditions.

Sample filtering was based on the \code{minfi} R package,
while probe filtering was based on the \code{ChAMP} R package.

Additionally, MethylCallR provides options to remove FLAG probes
and mapping failed probes.

FLAG probes have a low replication rate that does not meet the
manufacturer's quality control criteria (\code{FLAG}).
Mapping failed probes have an ambiguous position in the GRCh38/hg38 genome,
while positions are provided in GRCh37/hg19 (\code{hg38.mapfail}).

If these options are set to TRUE, MethylCallR will remove these probes
according to the manifest provided by Illumina and add them to the
\code{filtered.CpG.list}.

If your data uses the microarray version EPICv2, we recommend setting the
\code{RemoveDup} parameter to TRUE. Setting this parameter to TRUE will ensure
that duplicated probes in EPICv2 are selected to maintain characteristics
similar to those in the previous microarray version. For example,
if probe X was a type 1 probe in EPICv1, and both type 1 and type 2 probes
coexist in EPICv2, only the information related to the type 1 probe will be
preserved.

The relevant data (\code{Duplicated.Probes.preset}) were generated by
selecting probes with characteristics that do not match between two
versions (EPICv2 and EPICv1) of the Illumina manifest through comparison.
}
\examples{
\dontrun{
# To run filtering function with default setting.
data.filtered <- MeCall.filtering(data = data.Import, badSample.minfi = TRUE, 
badSampleCutoff.minfi = 10.5, detectionP = TRUE, detP.probe.cutoff = 0.01, 
detP.sample.cutoff = 0.1, detP.failed.ratio = 0, beadcount = TRUE, beadcutoff = 0.05, 
NoCG = TRUE, SNPmask = TRUE, population = "General", Multihit = TRUE, Sexchr = "XY", 
FLAG = TRUE, hg38.mapfail = TRUE)

# Or just
data.filtered <- MeCall.Filtering(data = data.Import)
}

}
\references{
Aryee, M.J. et al. (2014). Minfi: a flexible and comprehensive Bioconductor package for the
analysis of Infinium DNA methylation microarrays. Bioinformatics.
\url{https://doi.org/10.1093/bioinformatics/btu049}

Tian, Y. et al. (2017). ChAMP: updated methylation analysis pipeline for Illumina BeadChips.
Bioinformatics. \url{https://doi.org/10.1093/bioinformatics/btx513}

For Multihit probes (EPICv1 and 450K)
Nordlund, J. et al. (2013). Genome-wide signatures of differential DNA methylation in pediatric
acute lymphoblastic leukemia. Genome Biology. \url{https://doi.org/10.1186/gb-2013-14-9-r105}

For Multihit probes (EPICv2) and mismatch probes
Peters, T.J. et al. (2024). Characterisation and reproducibility of the HumanMethylationEPIC
v2.0 BeadChip for DNA methylation profiling. BMC Genomics.
\url{https://doi.org/10.1186/s12864-024-10027-5}

For SNP associated probes
Zhou, W. et al. (2017). Comprehensive characterization, annotation and innovative use of
Infinium DNA methylation BeadChip probes. Nucleic Acids Research.
\url{https://doi.org/10.1093/nar/gkw967}
}
\seealso{
\code{\link[minfi:getQC]{minfi::getQC()}}, \code{\link[ChAMP:champ.filter]{ChAMP::champ.filter()}}, \code{\link[=MeCall.ReadIdat]{MeCall.ReadIdat()}}
}
\author{
Hyun-Ho Yang
}
