% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.FindBatch.R
\name{MeCall.FindBatch}
\alias{MeCall.FindBatch}
\title{Finding potential confounding factors.}
\usage{
MeCall.FindBatch(meth, pd, batches = NULL, n.comp = 0)
}
\arguments{
\item{meth}{A matrix of methylation level.}

\item{pd}{A dataframe of sample data file is contained in an object returned by \code{MeCall.ReadIdat()} or \code{MeCall.Filtering()}.}

\item{batches}{vector. The column names present in the sample data file. If set to NULL, the correlation coefficient will be calculated for all column names present in the sample data file.}

\item{n.comp}{Correlation coefficients are calculated using a given number of principal components. If set to 0, a maximum of 15 principal components are used.}
}
\value{
a ggplot2 object representing a heatmap that illustrates the
correlation between principal components of methylation level and variables
in sample data file. The correlation coefficients are displayed at the
center of each tile. Blue tiles indicate significant correlations
(p < 0.05), with darker shades of blue representing lower p-values.
}
\description{
Perform principal component analysis (PCA) on the input data and calculate
the correlations between the principal components and the variables
specified in the batch parameter to assess the association between these
variables and the input data.
}
\examples{
\dontrun{
data.Heatmap <- MeCall.FindBatch(meth = data.Norm$beta, pd = data.filtered$pd, batches=NULL, 
n.comp = 0)

# To call the ggplot2 object
data.Heatmap

# To save the ggplot2 object as pdf file
pdf("./your_directory/data_Heatmap.pdf")
data.Heatmap
dev.off()
}

}
\seealso{
\code{\link[=MeCall.ReadIdat]{MeCall.ReadIdat()}}, \code{\link[=MeCall.Filtering]{MeCall.Filtering()}}
}
\author{
Hyun-Ho Yang
}
