% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.MakeModelMatrix.R
\name{MeCall.MakeModelMatrix}
\alias{MeCall.MakeModelMatrix}
\title{Generate model matrix for differential methylation analysis.}
\usage{
MeCall.MakeModelMatrix(
  pd = data.rmBatch$modified.pd,
  covariate = c("Age", "Sex"),
  interest = "Sample_Group",
  PCA.result = data.PCA,
  n.pc = 2,
  Cell.comp = data.CellComp$cell.composition,
  celltypes = colnames(data.CellComp$cell.composition),
  adjustment = FALSE,
  meth = NULL
)
}
\arguments{
\item{pd}{A data frame of modified sample data file from \code{MeCall.RemoveBatch()}.}

\item{covariate}{Vector. The covariates identified by \code{MeCall.FindBatch()}.}

\item{interest}{The column names in the sample data file containing the phenotype of interest.}

\item{PCA.result}{A returned object from \code{MeCall.PCA()}. Set to \code{NULL} if you do not want to include it as covariate.}

\item{n.pc}{The number of principal components to include in the covariate.}

\item{Cell.comp}{A dataframe containing the estimated cell composition for each sample. Set to \code{NULL} if you do not want to include it as covariate.}

\item{celltypes}{A vector containing the names of the cell types to be included in the covariate from the given \code{Cell.comp} parameter.}

\item{adjustment}{Logical. Methylation levels are adjusted using linear regression for given covariates.}

\item{meth}{When \code{adjustment} is set to \code{TRUE}, the matrix of methylation levels to be adjusted.}
}
\value{
The design matrix including the given covariates.
}
\description{
This function generates a model matrix containing information about
covariates to be considered for conducting a differential methylation
analysis based on the provided data.

If the \code{adjustment} parameter is set to \code{TRUE}, the function will
perform linear regression on the methylation level matrix to correct
covariates instead of providing a model matrix. This function will
return a list object containing the adjusted methylation level matrix
and a data frame that includes covariate information.
}
\examples{
\dontrun{
# Run [MeCall.MakeModelMatrix] with default setting.
data.design <- MeCall.MakeModelMatrix(pd = data.rmBatch$modified.pd, 
covariate = c("Age","Sex"), interest = "Sample_Group", PCA.result = data.PCA, n.pc = 2, 
Cell.comp = data.CellComp$cell.composition, celltypes = colnames(data.CellComp$cell.composition),
adjustment=FALSE, meth=NULL)
}

}
\seealso{
\code{\link[=MeCall.RemoveBatch]{MeCall.RemoveBatch()}}, \code{\link[=MeCall.FindBatch]{MeCall.FindBatch()}}, \code{\link[=MeCall.PCA]{MeCall.PCA()}}, \code{\link[=MeCall.CellComp]{MeCall.CellComp()}}
}
\author{
Hyun-Ho Yang
}
