% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.Norm.R
\name{MeCall.Norm}
\alias{MeCall.Norm}
\title{Normalization of methylation level}
\usage{
MeCall.Norm(
  data = filtered,
  method = c("BMIQ", "Funn", "SWAN", "ENmix", "dasen", "Noob"),
  InterpolatedXY = FALSE,
  offset = 100
)
}
\arguments{
\item{data}{A return object from \code{MeCall.filtering()}.}

\item{method}{Normalization method ("BMIQ","Funn","SWAN","ENmix","dasen","Noob").}

\item{InterpolatedXY}{logical. If set to TRUE, the interpolatedXY method is activated, which normalizes CpG sites on sex chromosomes without bias based on independently normalized autosomal CpG sites. It is recommended to set this to TRUE if CpG sites on sex chromosomes are included in the methylation level data. This parameter applies only when the normalization method is selected as \code{Funn} or \code{dasen}.}

\item{offset}{Integer. A fixed value will be added when generating methylation levels. An integer value used identically to the value used in the \code{MeCall.ReadIdat()} function.}

\item{return.type}{The representation method of methylation levels (Beta-value = \code{B}, M-value = \code{M}).}
}
\value{
A matrix consisting of normalized methylation levels.
}
\description{
This function provide several method to normalize the data after
removing samples and probes of poor quality.

\code{MeCall.Norm()} function provides 6 normalization algorithms available
across 4 different R packages (ENmix, minfi, WateRmelon, and ChAMP) :
\itemize{
\item Functional normalization (\code{method = "Funn"})
\item Beta-mixture quantile normalization (BMIQ) (\code{method = "BMIQ"})
\item Subset-quantile within array normalization (SWAN) (\code{method = "SWAN"})
\item dasen (\code{method = "dasen"})
\item Exponential-Normal mixture signal intensity background correction (ENmix) (\code{method = "ENmix"})
\item Normal-exponential out-of-band normalization (Noob) (\code{method = "Noob"})
}

Due to variations in input formats required by each normalization algorithm,
we recommend to utilize the return object of \code{MethylCallR.filtering()} function.

The interpolatedXY step is an additional process that normalizes CpG sites on sex chromosomes without bias based on independently normalized autosomal CpG sites. This step can be utilized when the method parameter is set to \code{Funn} or \code{dasen}. For more details, refer to \code{\link[wateRmelon:adjustedFunnorm]{wateRmelon::adjustedFunnorm()}} and \code{\link[wateRmelon:adjustedDasen]{wateRmelon::adjustedDasen()}}.
}
\examples{
\dontrun{
# For [functional normalization] and [Beta-value]
data.Norm.FUNN <- MeCall.Norm(data = data.filtered, method = c("Funn"), offset=100) 

# For [BMIQ] and [M-value]
data.Norm.BMIQ <- MeCall.Norm(data = data.filtered, method = c("BMIQ"), offset=100)
}

}
\references{
For Functional normalization
Triche, T.J., Jr et al. (2013). Low-level processing of Illumina Infinium DNA Methylation
BeadArrays. Nucleic Acids Research. /url{ttps://doi.org/10.1093/nar/gkt090}

For SWAN
Maksimovic, J. et al. (2012). SWAN: Subset-quantile Within Array Normalization for Illumina
Infinium HumanMethylation450 BeadChips. Genome Biology. /url{https://doi.org/10.1186/gb-2012-13-6-r44}

For Noob
Triche, T.J., Jr et al. (2013). Low-level processing of Illumina Infinium DNA Methylation
BeadArrays. Nucleic Acids Research. /url{https://doi.org/10.1093/nar/gkt090}

For ENmix
Xu, Z. et al. (2016). ENmix: a novel background correction method for Illumina
HumanMethylation450 BeadChip. Nucleic Acids Research. /url{https://doi.org/10.1093/nar/gkv907}

For BMIQ
Teschendorff, A.E. et al. (2013). A beta-mixture quantile normalization method for correcting
probe design bias in Illumina Infinium 450 k DNA methylation data. Bioinformatics.
/url{https://doi.org/10.1093/bioinformatics/bts680}

For dasen
Pidsley, R. et al. (2013). A data-driven approach to preprocessing Illumina 450K methylation
array data. BMC Genomics. /url{https://doi.org/10.1186/1471-2164-14-293}

For interpolatedXY
Wang, Y. et al. (2022). InterpolatedXY: a two-step strategy to normalize DNA methylation
microarray data avoiding sex bias. Bioinformatics.
/url{https://doi.org/10.1093/bioinformatics/btac436}
}
\seealso{
\code{\link[minfi:preprocessFunnorm]{minfi::preprocessFunnorm()}}, \code{\link[minfi:preprocessNoob]{minfi::preprocessNoob()}}, \code{\link[minfi:preprocessSwan]{minfi::preprocessSWAN()}},
\code{\link[ChAMP:champ.norm]{ChAMP::champ.norm()}}, \code{\link[ENmix:preprocessENmix]{ENmix::preprocessENmix()}}, \code{\link[wateRmelon:dasen]{wateRmelon::dasen()}}, \code{\link[=MeCall.filtering]{MeCall.filtering()}},
\code{\link[wateRmelon:adjustedFunnorm]{wateRmelon::adjustedFunnorm()}}, \code{\link[wateRmelon:adjustedDasen]{wateRmelon::adjustedDasen()}}
}
\author{
Hyun-Ho Yang
}
