% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.Outlier.R
\name{MeCall.Outlier}
\alias{MeCall.Outlier}
\title{Finding outlier samples using Mahalanobis distance.}
\usage{
MeCall.Outlier(
  meth = norm,
  pd = filtered$pd,
  group = NULL,
  nrep = 1000,
  sample.fraction = 0.9,
  conf = 0.999,
  n.component = 2,
  plot = FALSE
)
}
\arguments{
\item{meth}{A matrix of methylation level.}

\item{pd}{A dataframe of sample data file.}

\item{group}{The samples are divided into groups based on the specified criteria, and outliers will be identified within each group.}

\item{nrep}{The number of iterations for outlier detection.}

\item{sample.fraction}{The proportion of subsamples for each iteration.}

\item{conf}{Samples with distance values beyond the given confidence interval are considered as outliers.}

\item{n.component}{The number of principal components to include in calculation of Mahalanobis distance.}

\item{plot}{Create a scatter plot using PC1 and PC2 to display outliers.}
}
\value{
The sample name of obvious outlier and suggestive outlier. (Optional) A ggplot2 object that displays outliers.
}
\description{
The Mahalanobis distance is calculated independently for each subgroup of
samples based on the elements in the \code{group} variable.

In each subgroup, the Mahalanobis distance is calculated \code{nrep} times
using \code{n.component} principal components of the subset of \code{meth}.

Since the Mahalanobis distance is sensitive to outliers, in each iteration,
a random sub-sample of \code{sample.fraction*100}\% is selected to calculate
the covariance. This covariance is then applied to the entire sample
for Mahalanobis distance calculation.

Samples with distance values outside the \code{conf} confidence interval of
the chi-squared distribution with 2 degrees of freedom are designated
as outliers. Samples identified as outliers in more than 20\% of the
total iterations are classified as "obvious outliers", while
those identified in between 5\% and 20\% of the iterations are classified
as "suggestive outliers". The detection rates are included in the
returned result object.
}
\examples{
\dontrun{
# Identify outliers with default settings and generate scatter plot
Outliers <- MeCall.Outlier(meth = data.Norm, pd = data.filtered$pd, group = "Sample_Group", 
nrep = 1000, sample.fraction = 0.9, conf = 0.999, n.component = 2, plot = TRUE)

# To call the ggplot2 object
Outliers$Plot
}

}
\author{
Hyun-Ho Yang
}
