% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.PCA.R
\name{MeCall.PCA}
\alias{MeCall.PCA}
\title{Perfome principal component analysis.}
\usage{
MeCall.PCA(meth = mycellcomp)
}
\arguments{
\item{meth}{A matrix of methylation level.}
}
\value{
"DimReductionlist" S4 class object consisted of :
\itemize{Components} {The rotated data according to the \code{rotation}. The column names represent the principal component names, while the row names represent the samples.}
\itemize{cum.var} {The cumulative information explained by each principal component across the entire dataset.}
\itemize{sdev} {the square roots of the eigenvalues of the covariance/correlation matrix}
\itemize{rotation} {A matrix containing the principal components that represent the eigenvectors.}
\itemize{center} {A logical value indicating whether the data has been zero-centered.}
\itemize{scale} {A logical value indicating whether the variance of the variables has been scaled to have unit variance.}
see \code{\link[stats:prcomp]{stats::prcomp()}} for detail
}
\description{
Returns a structured object containing the results of the
principal component analysis (PCA) and the cumulative variance.

The \code{DimReductionlist} class generated by this function can be passed to
the \code{MeCall.Plot()} function to create a scree plot or PCA plot of
the data using the selected principal components.

In addition, it can be passed to the \code{MeCall.MakeModelMatrix()} function
to include principal components as covariates before conducting a
differential methylation analysis.
}
\examples{
\dontrun{
# Run PCA.
data.PCA <- MeCall.PCA(meth = data.CellComp)
}

}
\seealso{
\code{\link[=MeCall.Plot]{MeCall.Plot()}}, \code{\link[=MeCall.MakeModelMatrix]{MeCall.MakeModelMatrix()}}
}
\author{
Hyun-Ho Yang
}
