% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.Plot.R
\name{MeCall.Plot}
\alias{MeCall.Plot}
\title{Generate various plots based on ggplot2 R package.}
\usage{
MeCall.Plot(
  data = data,
  type = c("scree", "pca", "Manhattan", "Volcano", "functional.region", "CGI.region"),
  pd = pd,
  batch.name = NULL,
  PC.x = 1,
  PC.y = 2,
  delta.cut = c(0, 0),
  Pval.cut = 0.05,
  highlight = NULL,
  highlight.color = NULL,
  gene.label = NULL
)
}
\arguments{
\item{data}{The \code{DimReductionlist} object from \code{MeCall.PCA()} for \code{scree} and \code{pca} plot. The \code{DMPresultlist} object from \code{MeCall.DMP()} for \code{Manhattan},\code{Volcano}, \code{functional.region}, and \code{CGI.region} plot.}

\item{type}{The type of plot to create.}

\item{pd}{A data frame of modified sample data file from \code{MeCall.RemoveBatch()}. (pca plot)}

\item{batch.name}{Highlight samples with different colors of dots based on the given variable. (pca plot)}

\item{PC.x}{The principal component to display on the x-axis. (pca plot)}

\item{PC.y}{The principal component to display on the y-axis.(pca plot)}

\item{delta.cut}{The threshold for significant delta-beta values. (Volcano, region plot)}

\item{Pval.cut}{The cutoff for the p-value regardless of the multiple testing correction method. (Manhattan, Volcano, region plot)}

\item{highlight}{A list object with named slots. Each slot contains a vector consisting of probe IDs that the user wants to highlight. (Manhattan plot)}

\item{highlight.color}{Highlight probes in each slot with dots of different specified colors. (Manhattan plot)}

\item{gene.label}{A vector consisting of probe IDs for which the user wants to indicate associated gene annotations. (Manhattan, Volcano plot)}
}
\value{
A ggplot2 object
}
\description{
This function generates the selected plot based on the input data
and returns a ggplot2 object.

For a \code{Manhattan} plot, users may want to submit a named list object
to the \code{highlight} parameter to simultaneously highlight probes
from multiple subgroups. Through this, users can freely designate
three or more levels of highlighting. To label probes with their
associated gene names, users can simply input a vector composed
of probe IDs into the \code{gene.label} argument.
}
\examples{
\dontrun{
# Run [MeCall.Plot] to generate scree plot.
Scree_plot <- MeCall.Plot(data = data.PCA, type= c("scree"))

# Run [MeCall.Plot] to generate pca plot.
pca_plot <- MeCall.Plot(data = data.PCA, type= c("pca"), batch.name="Sample_Group", 
PC.x = 1, PC.y = 2)

# Run [MeCall.Plot] to generate Manhattan plot.
Highlight.list = list(H1 = c("cg00000001","cg00000002"), H2 = c("cg00000003","cg00000004"), 
H3 = c("cg00000005","cg00000006"))
Manhattan_plot <- MeCall.Plot(data = data.DMPs, type= c("Manhattan"), Pval.cut = 0.05, 
highlight = Highlight.list, highlight.color = c("red","green"),gene.label = c("cg00000001",
"cg00000002","cg00000003"))

# Run [MeCall.Plot] to generate Volcano plot.
Volcano_plot <- MeCall.Plot(data = data.DMPs, type= c("Volcano"), Pval.cut = 0.05, 
delta.cut = c(-0.02,0.02), gene.label = c("cg00000001","cg00000002"))

# Run [MeCall.Plot] to generate functional/CGI region plot.
Region_plot <- MeCall.Plot(data = data.DMPs, type= c("functional.region"), Pval.cut = 0.05, 
delta.cut = c(-0.02,0.02))
}

}
\author{
Hyun-Ho Yang
}
