% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.Power.R
\name{MeCall.Power}
\alias{MeCall.Power}
\title{Simulate statistical power for EWAS.}
\usage{
MeCall.Power(
  DMP.object = DMPs,
  tissue.type = "Adult (PBMC)",
  pd = pd,
  interest = "Sample_Group",
  FDR = 0.05,
  DeltaBeta = NULL,
  save.plot = NULL
)
}
\arguments{
\item{DMP.object}{A returned object from \code{MeCall.DMP()} or a data frame with "adjusted.p" and "deltabeta" as column names and CpG id as row names. Please note that the results should include information for all probes (set to \code{cutoff.P = 1}).}

\item{tissue.type}{A tissue type of sample ("Adult (PBMC)", "Saliva", "Sperm", "Lymphoma","Placenta", "Liver", "Colon", "Blood adult", "Blood 5 year olds", "Blood newborns", "Cord-blood (whole blood)", and "Cord-blood (PBMC)"). See manual of pwrEWAS R package for detail.}

\item{pd}{A data frame of modified sample data file from \code{MeCall.RemoveBatch()}.}

\item{interest}{The column names in the sample data file containing the phenotype of interest.}

\item{FDR}{The false discovery rate (FDR) value applied when identifying DMPs.}

\item{DeltaBeta}{A vector containing delta beta values for which the user wants to determine the statistical power in a simulation.}

\item{save.plot}{The directory where the line graph generated by \code{pwrEWAS} will be saved. If set to \code{NULL}, the plot will not be created.}
}
\value{
A dataframe containing the statistical power of each delta beta value in the given sample size.
}
\description{
This function estimates the statistical power for each delta-beta of
methylation changes under given conditions using the \code{pwrEWAS} R package
to identify an appropriate delta-beta cutoff. If the \code{plot} argument is
set to \code{TRUE}, a \code{ggplot2} object representing the calculated statistical
power will also be returned.

Considering the statistical power and appropriate delta-beta cutoff is
important to minimize false positive results. Even with adequately large
sample sizes, we recommend considering probes with a methylation
difference of 2\% or more (|delta-beta| > 0.02) as statistically
significant due to the technical variability of the EPIC array.

Currently, the \code{pwrEWAS} R package is not available on Bioconductor, so
users interested in using this function will need to install the
\code{pwrEWAS} R package directly from GitHub : https://github.com/stefangraw/pwrEWAS
}
\examples{
\dontrun{
# Simulate statistical power and generate the power plot.
data.Power <- MeCall.Power(DMP.object = MeCall.DMP, tissue.type = "Adult (PBMC)", 
pd = MeCall.RemoveBatch$modified.pd, interest = "Sample_Group", FDR = 0.05, 
DeltaBeta = c(0.02, 0.05, 0.07, 0.1), plot=TRUE)
}

}
\references{
Graw, S. et al. (2019). pwrEWAS: a user-friendly tool for comprehensive power estimation for
epigenome wide association studies (EWAS). BMC Bioinformatics.
/url{https://doi.org/10.1186/s12859-019-2804-7}
}
\seealso{
\code{\link[pwrEWAS:pwrEWAS]{pwrEWAS::pwrEWAS()}}, \code{\link[=MeCall.DMP]{MeCall.DMP()}}
}
\author{
Hyun-Ho Yang
}
