% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.ReadIdat.R
\name{MeCall.ReadIdat}
\alias{MeCall.ReadIdat}
\title{Read the idat file included in the sample data file}
\usage{
MeCall.ReadIdat(
  idat.dir = NULL,
  pd.file = NULL,
  offset = 100,
  platform = c("450K", "EPICv1", "EPICv2")
)
}
\arguments{
\item{idat.dir}{The directory path where the idat file exists.}

\item{pd.file}{A file containing Sentrix_ID and Sentrix_Position (csv, txt, tsv).}

\item{offset}{Integer. A fixed value will be added when generating methylation levels.}

\item{platform}{The version of the microarray to import (450K, EPICv1, EPICv2)}
}
\value{
A list object containing basic informations to process methylation microarray :
\itemize{beta} {A beta-value matrix defined with columns as samples and rows as CpG sites}
\itemize{M} {A M-value matrix structured similarly to a beta-value matrix}
\itemize{intensity} {A total intensity matrix structured similarly to a beta-value matrix}
\itemize{meth} {A methylated intensity matrix structured similarly to a beta-value matrix}
\itemize{Unmeth} {An unmethylated intensity matrix structured similarly to a beta-value matrix}
\itemize{detP} {A detection P-value matrix structured similarly to a beta-value matrix}
\itemize{B.count} {A bead count matrix structured similarly to a beta-value matrix}
*\itemize{pd} {A data frame composed of sample metadata, which includes information about each sample}
\itemize{minfi.Set} {A vector consisting of strings that indicate the version of the microarray}
\itemize{beta} {A list object consisting of two minfi objects (RGChannelSet Set and MethylSet)}
}
\description{
\code{MeCall.ReadIdat()} will attempt to find IDAT files located
in the directory specified by \code{idat.dir} parameter by
combining the Sentrix ID and Sentrix position information
in the sample metadata (\code{pd.file} parameter).

If you want to load only some of the IDAT files present in the directory,
delete the information of unnecessary samples from the sample metadata
before running the function.
}
\examples{
\dontrun{
data.Import <- MeCall.ReadIdat(idat.dir = "~/IdatDir", pd.file = "~/IdatDir/Sample_Sheet.csv", 
offset=100, platform = c("EPICv1"))
}

}
\references{
Aryee, M.J. et al. (2014). Minfi: a flexible and comprehensive Bioconductor package for the
analysis of Infinium DNA methylation microarrays. Bioinformatics.
/url{https://doi.org/10.1093/bioinformatics/btu049}

Tian, Y. et al. (2017). ChAMP: updated methylation analysis pipeline for Illumina BeadChips.
Bioinformatics. /url{https://doi.org/10.1093/bioinformatics/btx513}
}
\seealso{
\code{\link[minfi:MethylSet-class]{minfi::MethylSet()}}, \code{\link[minfi:RGChannelSet-class]{minfi::RGChannelSet()}}, \code{\link[ChAMP:champ.import]{ChAMP::champ.import()}}
}
\author{
Hyun-Ho Yang
}
