% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeCall.SetChainFile.R
\name{MeCall.SetChainFile}
\alias{MeCall.SetChainFile}
\title{Creating a reference file to convert between different versions of Illumina methylation microarray.}
\usage{
MeCall.SetChainFile(remove.list = NULL, auto.remove = FALSE, global.env = TRUE)
}
\arguments{
\item{remove.list}{A data frame containing "AddressA_ID" and "AddressB_ID" as column names and probe ID as row names. A subset of the Illumina manifest is highly recommended.}

\item{auto.remove}{logical. If set to \code{TRUE}, probes in the duplicated probe sets that are not addressed in the user-provided file are automatically removed based on the pre-set file implemented in \code{MethylCallR}.}

\item{global.env}{logical. If set to \code{TRUE}, returm object will be imported into global environment of R. If set to \code{FALSE}, function will return a data frame object.}
}
\value{
A list object containing the user-specified list of duplicated probes for removal
and the address chain file generated based on this list.
}
\description{
This function aims to facilitate the conversion between different versions of Illumina methylation microarrays by generating a "chain file" that links the unique values of the same probes across different versions.

Especially for EPICv2 data, users can validate and designate the list of duplicated probes generated by the user.

Users can create a chain file by using pre-processed data included in \code{MethylCallR} or by providing custom data in a data frame object.
Then, \code{MethylCallR} will reference this file to convert versions of Illumina methylation microarrays.

If the user designates the created objects as global variables named \code{Address_chain_file} or \code{Duplicated.Probes.preset}, respectively, \code{MethylCallR} will automatically reference these files.

Ensure that the \code{Address_chain_file} and \code{Duplicated.Probes.preset} are not masked by an external package or data.
}
\examples{
\dontrun{
# with default settings.
MeCall.SetChainFile(remove.list = NULL, auto.remove = FALSE, global.env = TRUE)

# To save the chain file from the user's file.
userfile <- user.dataframe.objects
Address_chain_file <- MeCall.SetChainFile(remove.list = userfile, auto.remove = TRUE, 
global.env = FALSE)
}

}
\author{
Hyun-Ho Yang
}
