% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Modified_EPICv2_manifest_in_TJ.Peters_BMCgenomics_2024}
\alias{Modified_EPICv2_manifest_in_TJ.Peters_BMCgenomics_2024}
\title{Modified EPICv2 manifest by TJ Peters}
\format{
A data frame with 936166 rows and 79 columns
}
\usage{
data(Modified_EPICv2_manifest_in_TJ.Peters_BMCgenomics_2024)
}
\description{
This file is modified version of EPICv2 manifest provided by Peters.
This file can be found in supplementary materials of Peters' study
(doi : url{https://doi.org/10.1186/s12864-024-10027-5}).

This file is used for filtering probes and defining duplicated probe set in EPICv2.
}
\references{
Peters, T.J. et al. (2024). Characterisation and reproducibility of the HumanMethylationEPIC v2.0 BeadChip for DNA methylation profiling. BMC Genomics. url{https://doi.org/10.1186/s12864-024-10027-5}
}
\keyword{datasets}
