% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SeSaMe_manifest}
\alias{SeSaMe_manifest}
\alias{HM450.hg38.manifest}
\title{SeSaMe manifest file}
\format{
A data frame with 486427 rows and 27 columns
}
\usage{
data(HM450.hg38.manifest)
}
\description{
These files are the SeSaMe manifest file downloaded from
the github \url{https://zwdzwd.github.io/InfiniumAnnotation}
Since Illumina manifest of 450K is not provide the genomic position based
the hg38, this file is required for filtering mapping fail probes and
generating address chain file for 450K.
}
\references{
Zhou, W. et al. (2017). Comprehensive characterization, annotation and innovative use of Infinium DNA methylation BeadChip probes. Nucleic Acids Research. \url{https://doi.org/10.1093/nar/gkw967}
}
\keyword{datasets}
