% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SNPmask}
\alias{SNPmask}
\alias{SNPmask.EPICv2}
\alias{SNPmask.EPICv2,}
\alias{SNPmask.EPICv1,}
\alias{SNPmask.450K}
\alias{SNPmask.EPICv1}
\title{SNP related probe list}
\format{
A data frame with 82740 rows and 36 columns (EPICv2)

A data frame with 105555 rows and 36 columns (EPICv1)

A data frame with 485577 rows and 66 columns (450K)
}
\usage{
data(SNPmask.EPICv2)

data(SNPmask.EPICv1)

data(SNPmask.450K)
}
\description{
These files contain probe IDs where SNPs are located closely.
These files can download at the github pages of SeSaMe (\url{https://zwdzwd.github.io/InfiniumAnnotation}).

Each row represents a probe.
Documentation for each column can be found in "header description" of github pages of SeSaMe (\url{https://zwdzwd.github.io/InfiniumAnnotation}).

These files are used for filtering probes.
}
\references{
For 450K and EPICv1 -
Zhou, W. et al. (2017). Comprehensive characterization, annotation and innovative use of Infinium DNA methylation BeadChip probes. Nucleic Acids Research. \url{https://doi.org/10.1093/nar/gkw967}

For EPICv2 -
Kaur, D. et al. (2023). Comprehensive evaluation of the Infinium human MethylationEPIC v2 BeadChip. Epigenetics Communications. \url{https://doi.org/10.1186/s43682-023-00021-5}
}
\keyword{datasets}
